# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Resale AI System",
    "description": "Local-first resale automation pipeline",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sync/relist/{sku}": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Relist Item",
        "operationId": "relist_item_api_sync_relist__sku__post",
        "parameters": [
          {
            "name": "sku",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku"
            }
          },
          {
            "name": "price_adjustment",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": -0.1,
              "title": "Price Adjustment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/relist-all": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Relist All",
        "operationId": "relist_all_api_sync_relist_all_post",
        "parameters": [
          {
            "name": "skus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Skus"
            }
          },
          {
            "name": "e2e_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "E2E Only"
            }
          },
          {
            "name": "price_adjustment",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": -0.1,
              "title": "Price Adjustment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}